+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 - Ability to restrict machine reservations per plan
 - Ability to restrict machine availabilities per plan
+- Admins cannot select the date when creating a refund invoice anymore
+- Fix a bug: logical sequence of invoices references is broken, when using the store module or the payments schedules
+- Fix a bug: refund invoices may generate duplicates in invoices references
 - Fix a security issue: updated webpack to 5.76.0 to fix [CVE-2023-28154](https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2023-28154)
 - [TODO DEPLOY] `rails db:seed`
 - [TODO DEPLOY] `rails fablab:maintenance:rebuild_stylesheet`